+++ execution/execute_all.go
@@ -17,6 +17,9 @@ func allExprExecutor(e ast.AllExpr) (expressionExecutor, error) {
 
 		result := true
 		if err := data.RangeSlice(func(i int, item *model.Value) error {
+			restore := withKeyVar(options, model.NewIntValue(int64(i)))
+			defer restore()
+
 			v, err := ExecuteAST(ctx, e.Expr, item, options)
 			if err != nil {
 				return err