+++ execution/execute_count.go
@@ -17,6 +17,9 @@ func countExprExecutor(e ast.CountExpr) (expressionExecutor, error) {
 
 		var count int64
 		if err := data.RangeSlice(func(i int, item *model.Value) error {
+			restore := withKeyVar(options, model.NewIntValue(int64(i)))
+			defer restore()
+
 			v, err := ExecuteAST(ctx, e.Expr, item, options)
 			if err != nil {
 				return err